# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "latch admin API",
    "description": "API JSON de l'admin latch (session cookie same-origin).",
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/projects": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "GET /api/projects — liste tous les projets (sans PIN, invariant §9.2).",
        "operationId": "list",
        "responses": {
          "200": {
            "description": "Liste des projets (sans PIN)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectListItem"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié"
          }
        }
      },
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "POST /api/projects — crée un nouveau projet.",
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Projet créé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetail"
                }
              }
            }
          },
          "400": {
            "description": "Requête invalide"
          },
          "401": {
            "description": "Non authentifié"
          },
          "403": {
            "description": "Origin invalide (CSRF)"
          }
        }
      }
    },
    "/api/projects/{id}": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "GET /api/projects/{id} — détail d'un projet avec ses versions et son PIN.",
        "operationId": "detail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identifiant du projet",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Détail du projet (avec PIN)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetail"
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié"
          },
          "404": {
            "description": "Projet inconnu"
          }
        }
      },
      "put": {
        "tags": [
          "projects"
        ],
        "summary": "PUT /api/projects/{id} — met à jour le nom ou le brand_name d'un projet.",
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identifiant du projet",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Projet mis à jour",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetail"
                }
              }
            }
          },
          "400": {
            "description": "Requête invalide"
          },
          "401": {
            "description": "Non authentifié"
          },
          "403": {
            "description": "Origin invalide (CSRF)"
          },
          "404": {
            "description": "Projet inconnu"
          }
        }
      },
      "delete": {
        "tags": [
          "projects"
        ],
        "summary": "DELETE /api/projects/{id} — supprime un projet et ses versions.",
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identifiant du projet",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Projet supprimé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié"
          },
          "403": {
            "description": "Origin invalide (CSRF)"
          },
          "404": {
            "description": "Projet inconnu"
          }
        }
      }
    },
    "/api/projects/{id}/code": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "POST /api/projects/{id}/code — active le code d'accès avec le PIN fourni.",
        "operationId": "set_code",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identifiant du projet",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetCodeReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Code activé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetail"
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié"
          },
          "403": {
            "description": "Origin invalide (CSRF)"
          }
        }
      },
      "delete": {
        "tags": [
          "projects"
        ],
        "summary": "DELETE /api/projects/{id}/code — désactive le code d'accès (PIN effacé).",
        "operationId": "clear_code",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identifiant du projet",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Code désactivé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetail"
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié"
          },
          "403": {
            "description": "Origin invalide (CSRF)"
          }
        }
      }
    },
    "/api/projects/{id}/comments/messages/{cid}": {
      "delete": {
        "tags": [
          "versions"
        ],
        "summary": "DELETE /api/projects/{id}/comments/messages/{cid} — modération (vérifie l'appartenance au projet).",
        "operationId": "moderate_delete_comment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identifiant du projet",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "cid",
            "in": "path",
            "description": "Identifiant du message",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Message supprimé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié"
          },
          "403": {
            "description": "Origin invalide (CSRF)"
          },
          "404": {
            "description": "Message hors projet ou inconnu"
          }
        }
      }
    },
    "/api/projects/{id}/deploy": {
      "post": {
        "tags": [
          "versions"
        ],
        "summary": "POST /api/projects/{id}/deploy — déploie un nouveau HTML, crée une version.",
        "operationId": "deploy",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identifiant du projet",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeployReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Version déployée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeployResponse"
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié"
          },
          "403": {
            "description": "Origin invalide (CSRF)"
          }
        }
      }
    },
    "/api/projects/{id}/versions/{n}": {
      "delete": {
        "tags": [
          "versions"
        ],
        "summary": "DELETE /api/projects/{id}/versions/{n} — supprime une version NON active.",
        "operationId": "delete_version",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identifiant du projet",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "n",
            "in": "path",
            "description": "Numéro de version",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Version supprimée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Version active : suppression refusée"
          },
          "401": {
            "description": "Non authentifié"
          },
          "403": {
            "description": "Origin invalide (CSRF)"
          },
          "404": {
            "description": "Version inconnue"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateProjectReq": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "brand_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "code_enabled": {
            "type": "boolean"
          },
          "comments_enabled": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "pin": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "OkResponse": {
        "type": "object",
        "description": "Réponse générique « succès » (`{\"ok\": true}`).",
        "required": [
          "ok"
        ],
        "properties": {
          "ok": {
            "type": "boolean"
          }
        }
      },
...
      "ProjectListItem": {
        "type": "object",
        "description": "Item de liste — **sans PIN** (invariant §9.2 : structurellement absent).",
        "required": [
          "id",
          "slug",
          "name",
          "code_enabled",
          "version_count",
          "comments_enabled"
        ],
        "properties": {
          "active_version_n": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Numéro (`n`) de la version active, ou `None` si aucun déploiement.\nOn n'expose PAS `active_version_id` (PK interne trompeuse) dans la liste."
          },
          "brand_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "code_enabled": {
            "type": "boolean"
          },
          "comments_enabled": {
            "type": "boolean"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "version_count": {
            "type": "integer",
            "format": "int32",
            "description": "Nombre total de versions du projet."
          }
        }
      },
...
      "SetCodeReq": {
        "type": "object",
        "required": [
          "pin"
        ],
        "properties": {
          "pin": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}